edit debugging part of the script

+++ pbpy/main.py
@@ -14,9 +14,10 @@ def main():
     if ncaa_id != '':
         g = game.Game(ncaa_id)
         g.setup_game()
+        g.create_plays()
         output = g.execute_game()
         df = pd.DataFrame(output)
-        df.to_csv('data/output/debug/' + str(ncaa_id) + '.csv')
+        df.to_csv('data/debug/' + str(ncaa_id) + '.csv')
     else:
         day_input = input("Input start date in format 'MM-DD-YYYY': ")
         end_input = input("Input end date in format 'MM-DD-YYYY': ")
@@ -39,14 +40,15 @@ def main():
                 print ("Successfully created the directory %s " % path)
             games = scrape.get_scoreboard(date)
             errors = []
+            msg = []
             for i in range(0, len(games)):
                 pct = i/len(games)*100
                 print(str(round(pct)) + '%' + ' '*(1 if round(pct)>=10 else 2) + '|' + '-'*round(pct/2) + ' '*round(50-pct/2) + '|')
                 ncaa_id = scrape.get_id('https://stats.ncaa.org' + games.iloc[i]['link'])
                 if not os.path.isfile('data/output/' + date + '/' + ncaa_id + '.csv'):
+                    print("SCRAPING GAME " + str(ncaa_id))
                     try:
-                        games.at[i, 'ncaa_id'] = ncaa_id
-                        print(games.at[i, 'link'])
+                        games.at[i, 'ncaa_id'] = ncaa_id                        
                         g = game.Game(ncaa_id)
                         if not os.path.isfile('data/raw/' + date + '/' + ncaa_id + '.json'):
                             raw = g.setup_game()
@@ -64,10 +66,16 @@ def main():
                             errors.append(1)
                         else:
                             errors.append(0)
-                    except:
+                        msg.append('')
+                    except Exception as e:
                         errors.append(1)
+                        msg.append(e)
                 
             games['error'] = errors
-            games.to_csv('data/games/' + date + '.csv', index=False)
+            games['msg'] = msg
+            if not os.path.isfile('data/games/' + date + '.csv'):
+                games.to_csv('data/games/' + date + '.csv', index=False)
+            else:
+                games.to_csv('data/games/' + date + '.csv', index=False, mode='a', header=False)
             day = day + timedelta(days=1)
             # TODO: Check output score against scoreboard
